refactor model ops
