+++ uniflow/op/model/LM/lm_model.py
@@ -4,14 +4,14 @@ import copy
 import json
 from typing import Any, Dict, List
 
-from uniflow.op.model.abs_llm_processor import AbsLLMProcessor
+from uniflow.op.model.abs_model import AbsModel
 from uniflow.op.model.constants import ERROR, ERROR_CONTEXT, ERROR_LIST, RESPONSE
 from uniflow.op.prompt import Context
 
 OUTPUT_SCHEMA_GUIDE = "Ensure the response is in json."
 
 
-class LLMDataProcessor(AbsLLMProcessor):
+class LmModel(AbsModel):
     """
     Data processor for processing data with a LLM model server.
     It handles serialization and deserialization of data,
@@ -64,9 +64,9 @@ class LLMDataProcessor(AbsLLMProcessor):
         }
 
 
-class JsonFormattedDataProcessor(AbsLLMProcessor):
+class JsonLmModel(AbsModel):
     """
-    Extends the LLMDataProcessor Class to ensure the response is in json.
+    Extends the LmModel Class to ensure the response is in json.
     """
 
     def _serialize(self, data: List[Context]) -> List[str]:
