+++ uniflow/op/model/LM/lm_rater_model.py
@@ -12,11 +12,11 @@ from uniflow.op.model.constants import (
     SCORES,
     VOTES,
 )
-from uniflow.op.model.llm_processor import JsonFormattedDataProcessor, LLMDataProcessor
+from uniflow.op.model.LM.lm_model import JsonLmModel, LmModel
 from uniflow.op.prompt import PromptTemplate
 
 
-class LLMRater(LLMDataProcessor):
+class LmRaterModel(LmModel):
     """LLM Rater"""
 
     def __init__(
@@ -99,7 +99,7 @@ class LLMRater(LLMDataProcessor):
         return data
 
 
-class OpenAIJsonFormattedLLMRater(JsonFormattedDataProcessor):
+class OpenAIJsonFormattedLLMRater(JsonLmModel):
     """OpenAI formatted LLM Rater"""
 
     def __init__(
@@ -178,7 +178,7 @@ class OpenAIJsonFormattedLLMRater(JsonFormattedDataProcessor):
         return data
 
 
-class HuggingfaceJsonFormattedLLMRater(LLMDataProcessor):
+class HuggingfaceJsonFormattedLLMRater(LmModel):
     """Huggingface formatted Json Formatted LLM Rater"""
 
     def __init__(
